rename pub-card-text to pub-card-paragraph
