+++ publiq-ui/pub-card.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, Story, Canvas, ArgsTable } from '@storybook/addon-docs/blocks';
 import PubCard from './pub-card.vue';
-import PubCardText from './pub-card-text.vue';
+import PubCardParagraph from './pub-card-paragraph.vue';
 
 <Meta title="Components/pub-card" component={PubCard} />
 
@@ -15,9 +15,9 @@ export const TemplateCard = (args, { argTypes }) => {
 export const TemplateCardWithCardText = (args, { argTypes }) => {
   return {
     props: Object.keys(argTypes),
-    components: { PubCard, PubCardText },
+    components: { PubCard, PubCardParagraph },
     template:
-      '<pub-card><pub-card-text>Paragraph 1</pub-card-text><pub-card-text>Paragraph 2</pub-card-text><pub-card-text>Paragraph 3 (will have its bottom margin removed)</pub-card-text></pub-card>',
+      '<pub-card><pub-card-paragraph>Paragraph 1</pub-card-paragraph><pub-card-paragraph>Paragraph 2</pub-card-paragraph><pub-card-paragraph>Paragraph 3 (will have its bottom margin removed)</pub-card-paragraph></pub-card>',
   };
 };
 
@@ -39,11 +39,11 @@ export const TemplateCardWithCardText = (args, { argTypes }) => {
 
 ```javascript
 <pub-card>
-  <pub-card-text>Paragraph 1</pub-card-text>
-  <pub-card-text>Paragraph 2</pub-card-text>
-  <pub-card-text>
+  <pub-card-paragraph>Paragraph 1</pub-card-paragraph>
+  <pub-card-paragraph>Paragraph 2</pub-card-paragraph>
+  <pub-card-paragraph>
     Paragraph 3 (will have its bottom margin removed)
-  </pub-card-text>
+  </pub-card-paragraph>
 </pub-card>
 ```
